+++ src/js/checkout.js
@@ -1,6 +1,6 @@
 import { loadHeaderFooter } from "./utils.mjs";
 
 loadHeaderFooter().then(() => {
-    document.querySelector(".icon-cart").innerHTML =
-      localStorage.getItem("so-cart-quantity") || 0;
-  });
+  document.querySelector(".icon-cart").innerHTML =
+    localStorage.getItem("so-cart-quantity") || 0;
+});
