+++ src/product_pages/index.html
@@ -1,27 +1,24 @@
 <!doctype html>
 
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
 
-<head>
-  <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Sleep Outside | Cedar Ridge Rimrock 2-person tent</title>
 
-  <title>Sleep Outside | Cedar Ridge Rimrock 2-person tent</title>
+    <link rel="stylesheet" href="../css/style.css" />
 
-  <link rel="stylesheet" href="../css/style.css" />
+    <script src="../js/product.js" type="module"></script>
+    <!-- <script src="../js/main.js"></script> -->
+  </head>
 
-  <script src="../js/product.js" type="module"></script>
-  <!-- <script src="../js/main.js"></script> -->
-</head>
+  <body>
+    <header class="divider" id="dynamic-header"></header>
 
-<body>
-  <header class="divider" id="dynamic-header">
-  </header>
+    <main class="divider"></main>
 
-  <main class="divider"></main>
-
-  <footer id="dynamic-footer"></footer>
-</body>
-
+    <footer id="dynamic-footer"></footer>
+  </body>
 </html>
